refactor: 더보기 버튼 파일 이동
